[AN] Bottom Padding 계산 로직 개선 (#305)
* feat: 키보드 높이에 따른 하단 패딩 적용 확장 함수 추가

* fix: 키보드 올라올 때 화면 가리는 현상 수정
