+++ android/Bottari/presentation/src/main/AndroidManifest.xml
@@ -19,8 +19,7 @@
         <activity
             android:name=".view.home.HomeActivity"
             android:exported="false"
-            android:windowSoftInputMode="adjustNothing"
-            android:launchMode="singleTask"/>
+            android:launchMode="singleTask" />
 
         <activity
             android:name=".view.checklist.ChecklistActivity"
