+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/common/base/BaseFragment.kt
@@ -4,8 +4,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import androidx.viewbinding.ViewBinding
 import com.bottari.logger.BottariLogger
@@ -44,7 +42,6 @@ abstract class BaseFragment<VB : ViewBinding>(
     ) {
         super.onViewCreated(view, savedInstanceState)
         BottariLogger.lifecycle(javaClass.simpleName)
-        setupWindowInsets()
     }
 
     override fun onStart() {
@@ -98,30 +95,5 @@ abstract class BaseFragment<VB : ViewBinding>(
 
         if (!loadingDialog.isAdded) return
         loadingDialog.dismissAllowingStateLoss()
-    }
-
-    private fun setupWindowInsets() {
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { view, insets ->
-            val imeVisible = insets.isVisible(WindowInsetsCompat.Type.ime())
-            val bottomInset = calculateBottomInset(insets, imeVisible)
-            view.setPadding(0, 0, 0, bottomInset)
-            insets
-        }
-    }
-
-    private fun calculateBottomInset(
-        insets: WindowInsetsCompat,
-        imeVisible: Boolean,
-    ): Int {
-        if (!imeVisible) return DEFAULT_BOTTOM_INSET
-
-        val imeInset = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
-        val systemBarInset = insets.getInsets(WindowInsetsCompat.Type.systemBars()).bottom
-
-        return (imeInset - systemBarInset).coerceAtLeast(0)
-    }
-
-    companion object {
-        private const val DEFAULT_BOTTOM_INSET = 0
     }
 }
