+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/common/extension/ViewExtension.kt
@@ -1,6 +1,8 @@
 package com.bottari.presentation.common.extension
 
 import android.view.View
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.isVisible
 
 fun View.fadeIn(duration: Long = 200L) {
@@ -25,4 +27,25 @@ fun View.fadeOut(duration: Long = 200L) {
                 visibility = View.GONE
             }
     }
+}
+
+fun View.applyImeBottomPadding() {
+    ViewCompat.setOnApplyWindowInsetsListener(this) { view, insets ->
+        val imeVisible = insets.isVisible(WindowInsetsCompat.Type.ime())
+        val bottomInset = calculateBottomInset(insets, imeVisible)
+        view.setPadding(0, 0, 0, bottomInset)
+        insets
+    }
+}
+
+private fun calculateBottomInset(
+    insets: WindowInsetsCompat,
+    imeVisible: Boolean,
+): Int {
+    if (!imeVisible) return 0
+
+    val imeInset = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
+    val systemBarInset = insets.getInsets(WindowInsetsCompat.Type.systemBars()).bottom
+
+    return (imeInset - systemBarInset).coerceAtLeast(0)
 }
